# docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Cluster:
    description: Kubernetes 集群信息
    properties:
      cluster_api:
        description: 集群 API 地址
        example: https://api.example.com
        type: string
      cluster_network:
        allOf:
        - $ref: '#/definitions/models.Network'
        description: 集群网络配置
      cluster_region:
        description: 集群区域
        example: cn-east-1
        type: string
      cluster_status:
        description: 集群状态
        example: true
        type: boolean
      cluster_subnet:
        description: 集群子网
        example:
        - '["subnet-1"'
        - ' "subnet-2"]'
        items:
          type: string
        type: array
      cluster_type:
        description: 集群类型
        example: kubernetes
        type: string
      cluster_version:
        description: 集群版本
        example: v1.20.0
        type: string
      cluster_zone:
        description: 集群可用区
        example:
        - '["zone-a"'
        - ' "zone-b"]'
        items:
          type: string
        type: array
      cn_name:
        description: 集群中文名称
        example: 生产环境集群
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      kube_config:
        description: Kubernetes 配置文件内容
        example: |-
          apiVersion: v1
          kind: Config
          clusters:
          - cluster:
              server: https://api.example.com
        type: string
      name:
        description: 集群名称
        example: prod-cluster
        type: string
      updatedAt:
        type: string
    type: object
  models.Network:
    description: 集群网络配置信息
    properties:
      pod_cidr:
        description: Pod 网段
        example: 10.244.0.0/16
        type: string
      svc_cidr:
        description: Service 网段
        example: 10.96.0.0/12
        type: string
    type: object
  v1.Taint:
    properties:
      effect:
        allOf:
        - $ref: '#/definitions/v1.TaintEffect'
        description: |-
          Required. The effect of the taint on pods
          that do not tolerate the taint.
          Valid effects are NoSchedule, PreferNoSchedule and NoExecute.
      key:
        description: Required. The taint key to be applied to a node.
        type: string
      timeAdded:
        description: |-
          TimeAdded represents the time at which the taint was added.
          It is only written for NoExecute taints.
          +optional
        type: string
      value:
        description: |-
          The taint value corresponding to the taint key.
          +optional
        type: string
    type: object
  v1.TaintEffect:
    enum:
    - NoSchedule
    - PreferNoSchedule
    - NoExecute
    type: string
    x-enum-varnames:
    - TaintEffectNoSchedule
    - TaintEffectPreferNoSchedule
    - TaintEffectNoExecute
info:
  contact: {}
paths:
  /api/v1/clusters:
    get:
      consumes:
      - application/json
      description: 获取所有集群的列表，支持分页和筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 集群类型
        in: query
        name: cluster_type
        type: string
      - description: 区域
        in: query
        name: region
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取集群列表
      tags:
      - clusters
    post:
      consumes:
      - application/json
      description: 创建一个新的 Kubernetes 集群
      parameters:
      - description: 集群信息
        in: body
        name: cluster
        required: true
        schema:
          $ref: '#/definitions/models.Cluster'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            $ref: '#/definitions/models.Cluster'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建新的集群
      tags:
      - clusters
  /api/v1/clusters/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除指定的集群
      parameters:
      - description: 集群ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 集群不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除集群
      tags:
      - clusters
    get:
      consumes:
      - application/json
      description: 根据ID获取集群的详细信息
      parameters:
      - description: 集群ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/models.Cluster'
        "404":
          description: 集群不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取单个集群信息
      tags:
      - clusters
    put:
      consumes:
      - application/json
      description: 根据ID更新集群的详细信息
      parameters:
      - description: 集群ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新的集群信息
        in: body
        name: cluster
        required: true
        schema:
          $ref: '#/definitions/models.Cluster'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/models.Cluster'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 集群不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新集群信息
      tags:
      - clusters
  /api/v1/clusters/{id}/nodes:
    get:
      consumes:
      - application/json
      description: 通过 kubeconfig 获取集群的 WorkNode 节点状态
      parameters:
      - description: 集群ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 集群不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取集群节点状态
      tags:
      - clusters
  /api/v1/clusters/{id}/nodes/{nodeName}/labels:
    patch:
      consumes:
      - application/json
      description: 更新指定集群中特定节点的标签
      parameters:
      - description: 集群ID
        in: path
        name: id
        required: true
        type: integer
      - description: 节点名称
        in: path
        name: nodeName
        required: true
        type: string
      - description: 节点标签
        in: body
        name: labels
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: 更新成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 集群不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新节点标签
      tags:
      - clusters
  /api/v1/clusters/{id}/nodes/{nodeName}/labels/{labelKey}:
    delete:
      consumes:
      - application/json
      description: 删除指定集群中特定节点的指定标签
      parameters:
      - description: 集群ID
        in: path
        name: id
        required: true
        type: integer
      - description: 节点名称
        in: path
        name: nodeName
        required: true
        type: string
      - description: 标签键
        in: path
        name: labelKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 集群不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除节点标签
      tags:
      - clusters
  /api/v1/clusters/{id}/nodes/{nodeName}/taints:
    patch:
      consumes:
      - application/json
      description: 更新指定集群中特定节点的污点
      parameters:
      - description: 集群ID
        in: path
        name: id
        required: true
        type: integer
      - description: 节点名称
        in: path
        name: nodeName
        required: true
        type: string
      - description: 节点污点
        in: body
        name: taints
        required: true
        schema:
          items:
            $ref: '#/definitions/v1.Taint'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 集群不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新节点污点
      tags:
      - clusters
  /api/v1/clusters/{id}/nodes/{nodeName}/taints/{taintKey}:
    delete:
      consumes:
      - application/json
      description: 删除指定集群中特定节点的指定污点
      parameters:
      - description: 集群ID
        in: path
        name: id
        required: true
        type: integer
      - description: 节点名称
        in: path
        name: nodeName
        required: true
        type: string
      - description: 污点键
        in: path
        name: taintKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 集群不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除节点污点
      tags:
      - clusters
swagger: "2.0"